Refactor: Extract duplicate decodeMetadata and app window opening logic to shared utilities
- Created utils/appUtils.ts with shared decodeMetadata and openAppFrontend functions
- Removed duplicate decodeMetadata implementations from App.tsx, InstalledApps.tsx, and Contexts.tsx
- Removed duplicate app window opening logic from App.tsx and InstalledApps.tsx
- Ensures consistent behavior across all files and reduces maintenance burden
